Make tests aware tcms.bugs may not be installed
- don't share BugFactory with the rest of the project
- conditionally skip tcms.bugs.tests modules
- ignore flake8's `E402 module level import not at top of file`
- ignore pylint's wrong-import-position
- conditionally assert expected results
