+++ tcms/bugs/tests/test_models.py
@@ -1,10 +1,18 @@
-from django.conf import settings
-from django.template.loader import render_to_string
-from django.test import TestCase
-from django.utils.translation import gettext_lazy as _
-from mock import patch
+# pylint: disable=wrong-import-position
+import unittest
 
-from tcms.tests.factories import BugFactory, UserFactory
+from django.conf import settings
+
+if 'tcms.bugs.apps.AppConfig' not in settings.INSTALLED_APPS:
+    raise unittest.SkipTest('tcms.bugs is disabled')
+
+from django.template.loader import render_to_string     # noqa: E402
+from django.test import TestCase                        # noqa: E402
+from django.utils.translation import gettext_lazy as _  # noqa: E402
+from mock import patch                                  # noqa: E402
+
+from tcms.tests.factories import UserFactory            # noqa: E402
+from tcms.bugs.tests.factory import BugFactory          # noqa: E402
 
 
 class TestSendMailOnAssigneeChange(TestCase):
